+++ app/core/bigquery.py
