+++ app/repository/bigquery_user_repo.py
@@ -2,7 +2,7 @@ from typing import Optional
 from fastapi import HTTPException, status
 from google.cloud import bigquery
 
-class bqUser:
+class UserRepository:
     def __init__(self, client: bigquery.Client, project_id: str, dataset_name: str, table_name: str):
         self.client = client
         self.project_id = project_id
